moved sides to a ems variable  (july 9)

+++ r_data.h
@@ -43,7 +43,7 @@ int R_FlatNumForName (char* name);
 
 // Called by P_Ticker for switches and animations,
 // returns the texture number for the texture name.
-int R_TextureNumForName (char *name);
+short R_TextureNumForName (char *name);
 int R_CheckTextureNumForName (char *name);
 void R_EraseCompositeCache(short texnum);
